modified payment.py
